Fix and suppress some compiler warnings

+++ lib/driver/rcinput.cpp
@@ -670,7 +670,7 @@ public:
 
 	void removeAll(void)
 	{
-		int size = items.size();
+		size_t size = items.size();
 		for (itemlist::iterator it = items.begin(); it != items.end(); ++it)
 		{
 			delete *it;
